% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{handle_nse_name}
\alias{handle_nse_name}
\title{Handle NSE name}
\usage{
handle_nse_name(x, env = rlang::caller_env())
}
\arguments{
\item{x}{[\link{character} or \link{name} or \link{quosure}]}

\item{env}{[\link{environment}] Caller environment}
}
\description{
Handle NSE name
}
\examples{
data_tbl <- tibble::tribble(
    ~color,     ~a, ~b, ~c, ~d,
    "blue",      1,  2, TRUE, FALSE,
    "green",     6,  2, TRUE, FALSE,
    "purple",    3,  3, TRUE, FALSE,
    "red",       2,  3, TRUE, FALSE,
    "yellow",    5,  1, TRUE, FALSE
)
foo <- function(data, col) {
    col <- handle_nse_name(col)
    data \%>\% dplyr::select(!!col)
}

data_tbl \%>\% foo(color)
data_tbl \%>\% foo("color")
data_tbl \%>\% foo(rlang::quo(color))

tmp <- rlang::sym("color")
col <- rlang::enquo(tmp)
data_tbl \%>\% foo(col)
}
