% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{handle_nse_names}
\alias{handle_nse_names}
\title{Handle NSE names}
\usage{
handle_nse_names(..., env = rlang::caller_env())
}
\arguments{
\item{...}{[\link{character} or \link{name} or \link{quosure}]}

\item{env}{[\link{environment}] Caller environment}
}
\description{
Handle NSE names
}
\examples{
bar <- function(data, ...) {
    cols <- handle_nse_names(...)
    data \%>\% dplyr::select(!!!cols)
}
col_c <- rlang::sym("c")
col_d <- "d"
col_d_enquo <- rlang::enquo(col_d)

data_tbl \%>\% bar(b, rlang::quo(a), "color", col_d_enquo, col_c)
}
