% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nise-package.R
\docType{package}
\name{nise-package}
\alias{nise}
\alias{nise-package}
\title{nise: Making the use of NSE in nested function calls fun!}
\description{
Functionality to programmatically leverage NSE within nested function calls without having to worry about "type casting" the supplied argument values/types
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io}

}
\keyword{internal}
